requirements.txt
aiogram>=3.4,<4.0
pydantic>=2.6,<3.0
pydantic-settings>=2.2,<3.0
openai>=1.0,<2.0
gspread>=6.0,<7.0
google-auth>=2.0,<3.0
